navgujaratsamay/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pillow>=11.0.0
pandas==2.2.3
beautifulsoup4==4.12.2
google-generativeai==0.3.2
python-dotenv==1.0.0
psutil==5.9.6
selenium==4.15.2
numpy==1.26.4
